chore: close/thread tab in channel page
