+++ src/App.tsx
@@ -29,6 +29,7 @@ import InvitedWorkspacePage from "pages/InvitedWorkspace";
 import { oneSignalId } from "utils/helper";
 import SearchPage from "pages/Search";
 import TrashPage from "pages/Trash";
+import ThreadList from "pages/Channel/ThreadList";
 
 function App() {
   const dispatch = useAppDispatch();
@@ -126,7 +127,10 @@ function App() {
 
           <Route path="saved" element={<>saved</>} />
           <Route path="messages" element={<>messages</>} />
-          <Route path="ch/:channelId" element={<ChannelPage />} />
+          <Route path="ch/:channelId/" element={<ChannelPage />}>
+            <Route path="" element={<ThreadList />} />
+            <Route path="close" element={<ThreadList type="close" />} />
+          </Route>
           <Route path="ch/:channelId/t/:threadId" element={<ThreadPage />} />
           <Route
             path="ch/:channelId/te/:threadId"
